requirements.txt
RPi.GPIO
Adafruit_DHT
pandas
openpyxl
pyserial
pyaudio
numpy
os
time
datetime
serial
i2c-tools
python3-smbus